fix: properties should have their associated annotation parameters
Fixes #214 (sorta). Turns out that annotations are rather painful to
process as you cannot call a method that you got from the annotation
type directly on it, so a proper fix would be to pass the actual
annotation instead of the annotation type so that both are available
during the processing.