Use Elements helper to check if a class already exists
To decide if a class has to be generated or has to be skipped because it
already exists in the classpath, the processor uses
`AbstractBuilderProcessor#classExists(String)` to check for the class
existence.

Currently the method relies on the class being present in the classpath,
but that presupposes that the annotation processor classpath is the same
as the source classpath, which is not always the case.

Changes the method to rely instead on the Elements class helper to
inspect the source classpath.

Fixes sundrio/sundrio#279